Add renderer tests to CI, and make two conversion routes agree
test/renderer.js covers what no other suite can reach. Everything else runs
under plain node and would pass with the interface completely broken; two
things exist only in a real Electron renderer -- Chromium's image decoder,
which load borrows, and the DOM.

It drives the actual UI rather than calling lab.run() directly. That matters:
clicking Run is what refreshes the tiles, and the first draft called the API
directly, produced no tiles at all, and failed on a null lookup. A test that
skips the interface is not testing the interface.

Eighteen assertions: the bridge surface and absence of Node globals, load
decoding to exact values in both spaces, the colour-space refusal naming its
own fix, a refused command appending nothing to the log, every tile drawing
something, §6 display defaults following the data kind, zoom shared across
tiles, the histogram view, an error surfacing without throwing into the page,
and no console errors.

CI runs it on all three platforms. Linux needs xvfb: Electron cannot open a
window on a headless runner, and show: false is not the same as needing no
display.

One assertion failed, and the failure was worth having. gray(toLinear(load))
did not match gray(load as=linear) -- the same value by two routes, differing
by one f32 ULP on about half the byte values, because the decode LUT computed
in double while the toLinear kernel reads an f32 intermediate. Numerically
nothing. For a lab that compares content hashes it is the difference between
two provenance chains agreeing and not.

Fixed by narrowing to f32 inside the LUT before applying the transfer
function, so the two routes are identical by construction rather than by
luck. Pinned by a test under plain node as well as in the renderer, and
recorded as determinism rule 4 in design-lab-model.md §5: where two routes
reach the same value, make them agree on purpose. Expect more of these.

Also extracted the PNG encoder into scripts/png.js, shared by make-sample.js
and the renderer test, which needs a fixture whose exact pixels it can assert
on after a round trip through Chromium.

And added the sRGB Wikipedia link to the glossary entry, as requested.

+++ native/addon_buffer.c
@@ -289,8 +289,20 @@ static napi_value BufferWrite(napi_env env, napi_callback_info info) {
 /* and no transcendental in the inner loop.                             */
 /* ------------------------------------------------------------------ */
 
+/*
+ * Note the deliberate round through float: the value is narrowed to f32
+ * BEFORE the transfer function is applied.
+ *
+ * Without it, `load(as=linear)` and `toLinear(load(...))` disagree by one f32
+ * ULP on about half the byte values -- mathematically the same result reached
+ * by two routes, differing because one of them stores an intermediate as f32
+ * and the other does not. Numerically that is nothing; for a lab that compares
+ * content hashes it is the difference between two provenance chains agreeing
+ * and not. Where two routes to the same value exist, make them agree on
+ * purpose.
+ */
 static float srgb_to_linear_byte(int i) {
-  const double s = (double)i / 255.0;
+  const double s = (double)(float)((double)i / 255.0);
   return (float)(s <= 0.04045 ? s / 12.92 : pow((s + 0.055) / 1.055, 2.4));
 }
